requirements.txt
chainlit
openai
mem0ai
python-dotenv
pandas
numpy 